Update design system for media upload edge case.
