+++ packages/story-editor/src/app/highlights/quickActions/useQuickActions.js
@@ -18,10 +18,10 @@
  * External dependencies
  */
 import { useCallback, useMemo, useRef } from '@web-stories-wp/react';
-import { __ } from '@web-stories-wp/i18n';
+import { __, sprintf, translateToExclusiveList } from '@web-stories-wp/i18n';
 import { useSnackbar, PLACEMENT, Icons } from '@web-stories-wp/design-system';
 import { trackEvent } from '@web-stories-wp/tracking';
-import { canTranscodeResource } from '@web-stories-wp/media';
+import { canTranscodeResource, resourceList } from '@web-stories-wp/media';
 
 /**
  * Internal dependencies
@@ -31,12 +31,14 @@ import updateProperties from '../../../components/inspector/design/updatePropert
 import useVideoTrim from '../../../components/videoTrim/useVideoTrim';
 import { useHistory } from '../../history';
 import { useConfig } from '../../config';
+import { useLocalMedia } from '../../media';
 import { ELEMENT_TYPES } from '../../../elements';
 import { useStory, useStoryTriggersDispatch, STORY_EVENTS } from '../../story';
 import useApplyTextAutoStyle from '../../../utils/useApplyTextAutoStyle';
-import { useMediaUpload } from '../../mediaUpload';
+import useFFmpeg from '../../media/utils/useFFmpeg';
 import { getResetProperties } from './utils';
 import { ACTIONS, RESET_PROPERTIES, RESET_DEFAULTS } from './constants';
+import MediaUploadWrapper from './mediaUploadWrapper';
 
 const {
   Bucket,
@@ -65,6 +67,12 @@ const {
  */
 const useQuickActions = () => {
   const {
+    allowedTranscodableMimeTypes,
+    allowedFileTypes,
+    allowedMimeTypes: {
+      image: allowedImageMimeTypes,
+      video: allowedVideoMimeTypes,
+    },
     capabilities: { hasUploadMediaAction },
     isRTL,
   } = useConfig();
@@ -98,7 +106,141 @@ const useQuickActions = () => {
       toggleTrimMode,
     })
   );
-  const { onOpenMediaUpload } = useMediaUpload();
+  const { resetWithFetch, updateVideoIsMuted, optimizeVideo, optimizeGif } =
+    useLocalMedia(
+      ({
+        actions: {
+          resetWithFetch,
+          updateVideoIsMuted,
+          optimizeVideo,
+          optimizeGif,
+        },
+      }) => {
+        return {
+          resetWithFetch,
+          updateVideoIsMuted,
+          optimizeVideo,
+          optimizeGif,
+        };
+      }
+    );
+
+  const { isTranscodingEnabled } = useFFmpeg();
+
+  // Media Upload Props
+  const allowedMimeTypes = useMemo(() => {
+    if (isTranscodingEnabled) {
+      return [
+        ...allowedTranscodableMimeTypes,
+        ...allowedImageMimeTypes,
+        ...allowedVideoMimeTypes,
+      ];
+    }
+    return [...allowedImageMimeTypes, ...allowedVideoMimeTypes];
+  }, [
+    allowedImageMimeTypes,
+    allowedVideoMimeTypes,
+    isTranscodingEnabled,
+    allowedTranscodableMimeTypes,
+  ]);
+
+  const transcodableMimeTypes = useMemo(() => {
+    return allowedTranscodableMimeTypes.filter(
+      (x) => !allowedVideoMimeTypes.includes(x)
+    );
+  }, [allowedTranscodableMimeTypes, allowedVideoMimeTypes]);
+
+  let onSelectErrorMessage = __(
+    'No file types are currently supported.',
+    'web-stories'
+  );
+  if (allowedFileTypes.length) {
+    onSelectErrorMessage = sprintf(
+      /* translators: %s: list of allowed file types. */
+      __('Please choose only %s to insert into page.', 'web-stories'),
+      translateToExclusiveList(allowedFileTypes)
+    );
+  }
+
+  /**
+   * Insert element such image, video and audio into the editor.
+   *
+   * @param {Object} resource Resource object
+   * @param {string} thumbnailURL The thumbnail's url
+   * @return {null|*} Return onInsert or null.
+   */
+  const insertMediaElement = useCallback(
+    (resource, thumbnailURL) => {
+      resourceList.set(resource.id, {
+        url: thumbnailURL,
+        type: 'cached',
+      });
+      updateElementsById({
+        elementIds: [selectedElements?.[0]?.id],
+        properties: { resource },
+      });
+    },
+    [selectedElements, updateElementsById]
+  );
+
+  const handleMediaSelect = useCallback(
+    (resource) => {
+      try {
+        if (isTranscodingEnabled && canTranscodeResource(resource)) {
+          if (transcodableMimeTypes.includes(resource.mimeType)) {
+            optimizeVideo({ resource });
+          }
+
+          if (resource.mimeType === 'image/gif') {
+            optimizeGif({ resource });
+          }
+        }
+        // WordPress media picker event, sizes.medium.source_url is the smallest image
+        insertMediaElement(
+          resource,
+          resource.sizes?.medium?.source_url || resource.src
+        );
+
+        if (
+          !resource.local &&
+          allowedVideoMimeTypes.includes(resource.mimeType) &&
+          resource.isMuted === null
+        ) {
+          updateVideoIsMuted(resource.id, resource.src);
+        }
+      } catch (e) {
+        showSnackbar({
+          message: e.message,
+          dismissable: true,
+        });
+      }
+    },
+    [
+      allowedVideoMimeTypes,
+      insertMediaElement,
+      isTranscodingEnabled,
+      optimizeGif,
+      optimizeVideo,
+      showSnackbar,
+      transcodableMimeTypes,
+      updateVideoIsMuted,
+    ]
+  );
+
+  const MediaPicker = useCallback(
+    (props) => (
+      <MediaUploadWrapper
+        title={__('Replace media', 'web-stories')}
+        buttonInsertText={__('Insert media', 'web-stories')}
+        onSelect={handleMediaSelect}
+        onClose={resetWithFetch}
+        type={allowedMimeTypes}
+        onSelectErrorMessage={onSelectErrorMessage}
+        {...props}
+      />
+    ),
+    [allowedMimeTypes, handleMediaSelect, resetWithFetch, onSelectErrorMessage]
+  );
 
   const undoRef = useRef(undo);
   undoRef.current = undo;
@@ -382,15 +524,15 @@ const useQuickActions = () => {
       actions.push({
         Icon: PictureSwap,
         label: ACTIONS.REPLACE_MEDIA.text,
-        onClick: (ev) => {
+        onClick: () => {
           dispatchStoryEvent(STORY_EVENTS.onReplaceForegroundMedia);
-          onOpenMediaUpload(ev);
 
           trackEvent('quick_action', {
             name: ACTIONS.REPLACE_MEDIA.trackingEventName,
             element: selectedElement?.type,
           });
         },
+        Wrapper: MediaPicker,
         ...actionMenuProps,
       });
     }
@@ -401,7 +543,7 @@ const useQuickActions = () => {
     foregroundCommonActions,
     hasUploadMediaAction,
     dispatchStoryEvent,
-    onOpenMediaUpload,
+    MediaPicker,
     selectedElement?.type,
   ]);
 
@@ -576,9 +718,8 @@ const useQuickActions = () => {
       baseActions.unshift({
         Icon: PictureSwap,
         label: ACTIONS.REPLACE_BACKGROUND_MEDIA.text,
-        onClick: (ev) => {
+        onClick: () => {
           dispatchStoryEvent(STORY_EVENTS.onReplaceBackgroundMedia);
-          onOpenMediaUpload(ev);
 
           trackEvent('quick_action', {
             name: ACTIONS.REPLACE_BACKGROUND_MEDIA.trackingEventName,
@@ -586,6 +727,7 @@ const useQuickActions = () => {
             isBackground: true,
           });
         },
+        Wrapper: MediaPicker,
         ...actionMenuProps,
       });
     }
@@ -617,7 +759,7 @@ const useQuickActions = () => {
     handleElementReset,
     handleFocusAnimationPanel,
     hasUploadMediaAction,
-    onOpenMediaUpload,
+    MediaPicker,
     resetProperties,
     selectedElement,
     showClearAction,
